ingestion/requirements.txt
# Google Drive API
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# OCR
google-cloud-vision==3.7.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# OpenAI
openai==1.12.0

# TypeSense
typesense==0.21.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
httpx==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
